requirements.txt
# Core
torch>=2.6.0 --index-url https://download.pytorch.org/whl/cu121
transformers>=4.52,<4.53
accelerate>=1.7.0
sentence-transformers>=2.7.0
datasets>=2.19.0

# Retrieval / similarity
faiss-cpu==1.7.4        # GPU 사용 시 faiss-gpu==1.7.4
faiss-gpu; platform_system!="Windows"

# LLM serving / quant
vllm==0.8.5.post1
safetensors>=0.4.2

# Classic ML & utils
scikit-learn>=1.5.0
numpy>=1.26,<2.0
pandas>=2.2.0
tqdm>=4.66.0
pyyaml>=6.0.1

# Optional (Jupyter UX)
jupyterlab>=4.1
ipywidgets>=8.1
